more cleaning. simplified x-client folder
