+++ backend/src/bork-protocol/utils/tweeting/tweet.ts
@@ -1,5 +1,11 @@
-import type { Content, IAgentRuntime, Memory, UUID } from '@elizaos/core';
-import { stringToUuid } from '@elizaos/core';
+import {
+  type Content,
+  type IAgentRuntime,
+  type Memory,
+  type UUID,
+  elizaLogger,
+  stringToUuid,
+} from '@elizaos/core';
 import type { Tweet } from 'agent-twitter-client';
 import type { TwitterService } from 'src/bork-protocol/services/twitter-service';
 
@@ -158,4 +164,43 @@ function splitParagraph(paragraph: string, maxLength: number): string[] {
   }
 
   return chunks;
+}
+
+export async function sendTweetAndCreateMemory(
+  twitterService: TwitterService,
+  response: Content,
+  roomId: UUID,
+  agentId: UUID,
+  userId: UUID,
+  inReplyToId?: string,
+): Promise<Memory[]> {
+  const memories: Memory[] = [];
+  const text = response.text;
+
+  if (!text) {
+    elizaLogger.error('[Twitter Client] No text to send');
+    return memories;
+  }
+
+  try {
+    const tweet = await twitterService.sendTweet(text, inReplyToId);
+    const memory: Memory = {
+      id: stringToUuid(`${tweet.id}-${agentId}`),
+      agentId,
+      content: {
+        text: tweet.text,
+        source: 'twitter',
+        url: tweet.permanentUrl,
+        inReplyTo: inReplyToId ? stringToUuid(inReplyToId) : undefined,
+      },
+      createdAt: tweet.timestamp * 1000,
+      roomId,
+      userId,
+    };
+    memories.push(memory);
+  } catch (error) {
+    elizaLogger.error('[Twitter Client] Error sending tweet:', error);
+  }
+
+  return memories;
 }
